queue-worker/requirements.txt
motor==3.4.0
pymongo==4.7.2
redis==5.0.4
python-dotenv==1.0.1
